tests/wpt/web-platform-tests/encoding-detection/th-windows-874-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>th windows-874 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-874", 'Expected windows-874');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/th-windows-874-late.sub.html");
</script>
